feat(core): implement multi-phase e-commerce architecture (P0–P8), wishlist system, intelligent search, and account management
This commit introduces foundational and scalable architecture across the platform, covering core commerce domains from data integrity to user engagement, aligned with Phases 0–8 of the product roadmap.

---

PHASE 0 — Foundation Remediation

* Standardized all monetary handling using integer kobo via Money value object and Eloquent cast
* Refactored pricing, cart, tax, and order services to eliminate float arithmetic
* Introduced price history tracking with audit trail via observers
* Migrated product option structures from JSON to relational schema
* Added product specifications, brands, and structured attributes
* Centralized VAT configuration across system
* Unified money formatting across Blade views via helper/component

---

PHASE 1 — Inventory & Bulk Import

* Implemented stock reservation system to prevent overselling
* Added stock adjustment logging with admin attribution
* Enforced available stock checks at checkout and payment stages
* Built bulk product importer (CSV/Excel) with validation and upsert logic
* Added admin interfaces for stock management and import workflows

---

PHASE 2 — Identity & Addresses

* Implemented OTP-based authentication (email/phone) and password reset flows
* Built user address book with default address handling
* Enabled guest-to-user account transition post-checkout
* Developed wholesale onboarding with document upload and admin approval workflow
* Introduced tier-based wholesale access and pricing visibility

---

PHASE 3 — Logistics & Checkout

* Implemented delivery zone-based pricing (weight/volume aware)
* Added pickup locations and delivery method selection
* Built shipment lifecycle with stage tracking (confirmed → delivered)
* Developed multi-step checkout with improved UX and order summary
* Created admin dispatch workflows and shipment tracking integration

---

PHASE 4 — Payments

* Added manual bank transfer flow with proof upload and admin reconciliation
* Implemented Pay-on-Delivery eligibility rules
* Extended refund system to support partial refunds
* Added reconciliation reporting and wholesale proforma invoice support

---

PHASE 5 — Notifications

* Integrated SMS/WhatsApp notification abstraction layer
* Implemented stage-based order notifications with idempotent delivery
* Added support for restock and low-stock alerts

---

PHASE 6 — Engagement & Growth

* Implemented wishlist module:

  * AJAX-based add/remove interactions (no page reload)
  * Persistent state for guest and authenticated users
  * Storefront UI integration
* Built intelligent search system:

  * Real-time product search
  * Trending and recent searches
  * Dynamic suggestion UX
* Laid groundwork for reviews, social sharing, and marketing features

---

PHASE 7 — Promotions

* Integrated coupon system into cart and checkout
* Added promotional pricing with scheduling support
* Built bulk price update tooling with audit tracking

---

PHASE 8 — Admin & Storefront Enhancements

* Implemented admin and customer account settings dashboards
* Improved storefront UX (navigation, trust signals, performance optimizations)
* Added operational admin tools (product management, reporting foundations)

---

Impact

* Establishes a production-grade, scalable e-commerce architecture
* Improves data integrity (money, inventory, pricing)
* Enhances user experience (search, wishlist, checkout)
* Enables operational efficiency (admin tools, inventory control)
* Lays foundation for growth features (reviews, promotions, marketing)

This commit represents a major milestone toward GoBuy v1 readiness.

+++ database/migrations/2026_06_15_003513_create_orders_table.php
@@ -25,9 +25,10 @@ return new class extends Migration
             $table->string('status')->default('pending')->index();
             $table->string('payment_status')->default('unpaid')->index();
 
-            $table->decimal('subtotal', 12, 2);
-            $table->decimal('delivery_fee', 12, 2)->default(0);
-            $table->decimal('total', 12, 2);
+            // Money is stored as integer kobo.
+            $table->unsignedBigInteger('subtotal');
+            $table->unsignedBigInteger('delivery_fee')->default(0);
+            $table->unsignedBigInteger('total');
 
             $table->timestamp('placed_at')->nullable();
             $table->timestamps();
